IDentrix/requirements.txt
# requirements.txt
# Dependencies for Person Re-Identification project

# Core ML libraries
torch>=1.12
torchvision>=0.13
timm>=0.9.2

# Data and evaluation
scikit-learn>=1.0
numpy>=1.22
Pillow>=9.0
matplotlib>=3.5

# Web UI
streamlit>=1.25

# Optional (for reproducibility/debugging)
tqdm
